Collapse Context/AssemblyEntry and remove unnecessary flags
