+++ src/coreclr/binder/assemblybindercommon.cpp
@@ -118,7 +118,6 @@ namespace BINDER_SPACE
             IF_FAIL_GO(pAssembly->Init(pPEImage, /* fIsInTPA */ FALSE ));
 
             pBindResult->SetResult(pAssembly);
-            pBindResult->SetIsFirstRequest(TRUE);
 
         Exit:
             return hr;
@@ -1058,13 +1057,8 @@ namespace BINDER_SPACE
                 hr = S_OK; // FindInExecutionContext returns S_FALSE when not found - reset to S_OK
                 SAFE_NEW(pContextEntry, ContextEntry);
 
-                pContextEntry->SetIsInTPA(pBindResult->GetIsInTPA());
                 pContextEntry->SetAssemblyName(pBindResult->GetAssemblyName(), TRUE /* fAddRef */);
                 pContextEntry->SetAssembly(pBindResult->GetAssembly());
-                if (pBindResult->GetIsFirstRequest())
-                {
-                    pContextEntry->SetIsFirstRequest(TRUE);
-                }
 
                 ExecutionContext *pExecutionContext = pApplicationContext->GetExecutionContext();
                 pExecutionContext->Add(pContextEntry);
